move tests to dedicated directory
