+++ moqtail-core/tsconfig.test.json
@@ -1,0 +1,8 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "paths": {
+      "bytes": ["../bytes/src/index.ts"]
+    }
+  }
+}
